🚚 api: move serializer to src context
