+++ api/src/certification/session-management/infrastructure/serializers/certification-serializer.js
@@ -3,10 +3,10 @@ import _ from 'lodash';
 
 const { Serializer, Deserializer } = jsonapiSerializer;
 
-import { CertificationCourse } from '../../../../src/certification/shared/domain/models/CertificationCourse.js';
-import { NO_EXAMINER_COMMENT } from '../../../../src/certification/shared/domain/models/CertificationReport.js';
-import { WrongDateFormatError } from '../../../../src/shared/domain/errors.js';
-import { isValidDate } from '../../../../src/shared/infrastructure/utils/date-utils.js';
+import { WrongDateFormatError } from '../../../../../src/shared/domain/errors.js';
+import { isValidDate } from '../../../../shared/infrastructure/utils/date-utils.js';
+import { CertificationCourse } from '../../../shared/domain/models/CertificationCourse.js';
+import { NO_EXAMINER_COMMENT } from '../../../shared/domain/models/CertificationReport.js';
 
 const serializeFromCertificationCourse = function (certificationCourse) {
   return new Serializer('certifications', {
