+++ api/tests/certification/session-management/unit/infrastructure/serializers/certification-serializer_test.js
@@ -1,8 +1,8 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/certification-serializer.js';
-import { CertificationCourse } from '../../../../../src/certification/shared/domain/models/CertificationCourse.js';
-import { NO_EXAMINER_COMMENT } from '../../../../../src/certification/shared/domain/models/CertificationReport.js';
-import { WrongDateFormatError } from '../../../../../src/shared/domain/errors.js';
-import { domainBuilder, EMPTY_BLANK_AND_NULL, expect } from '../../../../test-helper.js';
+import { WrongDateFormatError } from '../../../../../../src/shared/domain/errors.js';
+import * as serializer from '../../../../../../src/certification/session-management/infrastructure/serializers/certification-serializer.js';
+import { CertificationCourse } from '../../../../../../src/certification/shared/domain/models/CertificationCourse.js';
+import { NO_EXAMINER_COMMENT } from '../../../../../../src/certification/shared/domain/models/CertificationReport.js';
+import { domainBuilder, EMPTY_BLANK_AND_NULL, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | certification-serializer', function () {
   describe('#deserialize', function () {
